+++ frontend/src/components/PlayerCastButton.vue
@@ -1,23 +1,26 @@
 <script setup lang="ts">
 import { cleanupCastPlayer, initializeCast } from '~/logic/castFunctions'
-import { chromecastAvailable } from '~/logic/castRefs'
+import { castPlayer, chromecastAvailable } from '~/logic/castRefs'
 import { debugLog } from '~/logic/logger'
+import { currentVolume } from '~/logic/volume'
 
 const castButton = useTemplateRef('castButton')
 
-function updateStyle(elem: HTMLElement) {
-  const shadow = elem.shadowRoot
-  if (!shadow)
-    return
-  const styleNode = shadow.querySelector('style')
-  if (styleNode) {
-    styleNode.textContent = '.cast_caf_state_c {fill: hsl(32 100% 50%);}.cast_caf_state_d {fill: var(--disconnected-color, #7d7d7d);}.cast_caf_state_h {opacity: 0;}'
+watchEffect(() => {
+  if (castButton.value && castButton.value.shadowRoot) {
+    const shadow = castButton.value.shadowRoot
+    if (!shadow)
+      return
+    const styleNode = shadow.querySelector('style')
+    if (styleNode) {
+      styleNode.textContent = '.cast_caf_state_c {fill: hsl(32 100% 50%);}.cast_caf_state_d {fill: var(--disconnected-color, #7d7d7d);}.cast_caf_state_h {opacity: 0;}'
+    }
   }
-}
+})
 
-watch(castButton, (newVal) => {
-  if (newVal && newVal.shadowRoot) {
-    updateStyle(newVal)
+watchEffect(() => {
+  if (castPlayer.value && castPlayer.value.volumeLevel !== currentVolume.value) {
+    castPlayer.value.volumeLevel = currentVolume.value
   }
 })
